feat: add default values to edit user form
