+++ src/UI/Input/Input.tsx
@@ -1,42 +1,42 @@
-import React, { forwardRef, PropsWithChildren, SyntheticEvent } from 'react';
-import { FieldErrors } from 'react-hook-form';
+import React, { PropsWithChildren } from 'react';
+import { FieldErrors, Path, UseFormRegister, RegisterOptions } from 'react-hook-form';
 import styles from './styles.module.scss';
 import FormError from '../FormError';
+import { UserDto } from '../../models/User/UserDto';
 
 interface Props {
-  name: string;
+  name: Path<UserDto>;
+  register: UseFormRegister<UserDto>;
+  options?: RegisterOptions;
   label: string;
   isVisible?: boolean;
   errors: FieldErrors;
-  onChange: (e: SyntheticEvent) => void;
-  onBlur: (e: SyntheticEvent) => void;
 }
 
-const Input = forwardRef<HTMLInputElement, PropsWithChildren<Props>>(
-  (
-    { name, label, isVisible = true, errors, children, onChange, onBlur }: PropsWithChildren<Props>,
-    ref,
-  ) => {
-    return (
-      <div className={styles.inputWrapper}>
-        <label className={styles.label} htmlFor={name}>
-          {label}
-        </label>
-        <input
-          id={name}
-          type={isVisible ? 'text' : 'password'}
-          className={styles.input}
-          ref={ref}
-          onChange={onChange}
-          onBlur={onBlur}
-        />
-        {children}
-        <FormError errors={errors} name={name} />
-      </div>
-    );
-  },
-);
-
-Input.displayName = 'Input';
+const Input = ({
+  name,
+  register,
+  options = {},
+  label,
+  isVisible = true,
+  errors,
+  children,
+}: PropsWithChildren<Props>) => {
+  return (
+    <div className={styles.inputWrapper}>
+      <label className={styles.label} htmlFor={name}>
+        {label}
+      </label>
+      <input
+        id={name}
+        {...register(name, options)}
+        type={isVisible ? 'text' : 'password'}
+        className={styles.input}
+      />
+      {children}
+      <FormError errors={errors} name={name} />
+    </div>
+  );
+};
 
 export default Input;
