+++ src/components/EditUserForm/styles.module.scss
@@ -9,6 +9,19 @@
   }
 }
 
+.form {
+  width: 100%;
+}
+
+.button {
+  width: 100%;
+}
+
+.roles {
+  display: flex;
+  gap: 10px;
+}
+
 .closeButton {
   position: absolute;
   right: -5px;
