--- src/components/UserForm/index.ts
@@ -1,1 +1,0 @@
-export { default } from './UserForm';
