--- src/components/UserForm/styles.module.scss
@@ -1,15 +1,0 @@
-@use '../../styles/mixins/breakpoints.scss' as bp;
-@import '../../styles/utils/variables';
-
-.form {
-  width: 100%;
-}
-
-.button {
-  width: 100%;
-}
-
-.roles {
-  display: flex;
-  gap: 10px;
-}
